refactor(auth): renombrar carpetas dto/Request y dto/Response a minúsculas y actualizar imports
