--- src/modules/auth/application/dto/Response/RefreshTokenDto.ts
@@ -1,3 +1,0 @@
-export interface RefreshTokenDto {
-  refreshToken: string;
-}
